+++ engines/zxing-cpp-wasm/src/index.ts
@@ -2,6 +2,7 @@ export { supportsWasmSimd, supportsWebAssembly, detectRuntimeKind } from "./capa
 export { ZxingCppWasmError, ZxingCppWasmLoader } from "./loader.js";
 export { ZxingCppWasmEngineImpl } from "./engine.js";
 export { ZXING_CPP_WASM_BUILD_METADATA } from "./metadata.js";
+export { SCANLY_TO_NATIVE_FORMAT, NATIVE_TO_SCANLY_FORMAT, nativeFormatsFor, scanlyFormatFromNative } from "./formats.js";
 export type * from "./types.js";
 
 import { ZxingCppWasmEngineImpl } from "./engine.js";